requirements.txt
Flask
flasgger
python-dotenv
cohere
prometheus_flask_exporter
prometheus_client
pandas